requirements.txt
pandas==2.2.2
openpyxl==3.1.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pymysql==1.1.1
python-dotenv==1.0.1
pyarrow==17.0.0
numpy==2.1.3
